requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
PyMySQL>=1.1
pymongo>=4.6
python-dotenv>=1.0
